+++ imports/plugins/core/taxes/client/settings/custom.html
@@ -13,7 +13,7 @@
     }}
   </div>
 
-  <div class="add-new-tax-rate-button">
+  <div class="add-new-tax-rate-button pull-right">
     {{> React editButton }}
   </div>
