+++ imports/plugins/core/ui-grid/client/index.js
@@ -1,1 +1,1 @@
-import "./griddle.jsx";
+import "./griddle.js";
